requirements.txt
Flask
numpy
pytube
opencv-python
matplotlib
azure-ai-anomalydetector